Decouple templates and authorization fetching/application
